+++ lib/whois/record/parser/whois.markmonitor.com.rb
@@ -21,8 +21,15 @@ module Whois
       #
       class WhoisMarkmonitorCom < BaseIcannCompliant
         self.scanner = Scanners::BaseIcannCompliant, {
-            pattern_available: /^No match for/
+            pattern_available: /^No match for/,
+            pattern_throttled: /^You have exceeded your quota of queries\./,
         }
+
+
+        def response_throttled?
+          !!node("response:throttled")
+        end
+
       end
 
     end