+++ lib/whois/record/scanners/base_icann_compliant.rb
@@ -20,6 +20,7 @@ module Whois
         self.tokenizers += [
             :skip_head,
             :scan_available,
+            :scan_throttled,
             :skip_empty_line,
             :skip_blank_line,
             :scan_keyvalue,
@@ -29,6 +30,12 @@ module Whois
         tokenizer :scan_available do
           if settings[:pattern_available] && @input.skip_until(settings[:pattern_available])
             @ast['status:available'] = true
+          end
+        end
+
+        tokenizer :scan_throttled do
+          if settings[:pattern_throttled] && @input.skip_until(settings[:pattern_throttled])
+            @ast['response:throttled'] = true
           end
         end